Develop a system for easily defining autons: see below
In choreo, there's a feature called path splits.
Define a full auto path in choreo, and then split it at each score and intake, alternating.

Programatically compose a command adding the scores and intakes between every other path split (see AutoBuilder.cpp)

Also add a simple auto that drives forward and scores blindly, ignores alliance color

+++ src/main/cpp/AutoBuilder.cpp
@@ -21,7 +21,7 @@ frc2::CommandPtr AutoBuilder::AutoScore(frc::Pose2d scoring_pose) {
 
 frc2::CommandPtr AutoBuilder::MultiL4Auto(trajectory_t path) {
   std::vector<frc2::CommandPtr> commands;
-  for (int splitInd = 0; (unsigned)splitInd < path.splits.size(); splitInd++) {
+  for (unsigned splitInd = 0; splitInd < path.splits.size(); splitInd++) {
     if (const auto segment = path.GetSplit(splitInd)) {
       if (splitInd % 2 == 0) { // even splits are driving to the reef
         const auto scoring_pose = segment.value().GetFinalPose(isRed()).value();
